requirements.txt
gunicorn==20.1.0
Django==4.1
djangorestframework==3.14.0
django-environ==0.9.0
requests==2.28.1
responses==0.21.0
whitenoise==6.3.0
django-mysql==4.8.0